updated code. updated readme
