+++ challenges/tree_intersection/README.md
@@ -1,0 +1,28 @@
+# Tree Intersection
+**Author** : Jay Adams
+**Version**: 0.2.0
+
+## Overview
+
+
+
+## Spec
+---------------
+Write a function that accepts two binary trees and returns a set of the intersection.
+
+
+## Assets
+![pseudocode](../../assets/tree_intersection.jpg)
+
+
+
+## Architechture
+Python 3.6.4
+ENV
+Pytest
+
+
+## API
+None
+
+## Change log
